refactor: move ProjectType and ProjectOptions into src/types.ts
Both types are domain language for the generator (the user-facing
shape of "what to generate"), not utilities. Moving them out of
src/utils/validation.ts and into a top-level src/types.ts:

  - removes the implicit "validation owns the canonical type"
    coupling — validation now imports the type, doesn't define it
  - gives ProjectOptions a neutral home (was being defined inside
    src/cli/resolveProjectOptions.ts after the previous step's
    collapse, which made the resolver appear to own a domain type
    that's also consumed by the generator)
  - sets up the next commit, where validation.ts is mostly emptied
    and the remaining file deletions become straightforward

Importer updates (mechanical):
  - src/cli/generateProject.ts:    ProjectOptions + ProjectType from '../types'
  - src/cli/resolveProjectOptions.ts: same
  - src/cli/options-parser.ts:     ProjectType from '../types'
  - tests/resolve-project-options.test.ts: ProjectType from '../src/types'

src/utils/validation.ts now imports `ProjectType` from '../types' so
isValidProjectType's type predicate still resolves; everything else
in that file is unchanged.

Verified: bun test 40 pass / 4 fail (same pre-existing failures).
