+++ src/types.ts
@@ -1,0 +1,6 @@
+export type ProjectType = 'fullstack' | 'client-only' | 'library' | 'vanilla'
+
+export type ProjectOptions = {
+  name: string
+  type: ProjectType
+}
